+++ frontend/public/e-uscito-joypad.css
@@ -88,18 +88,6 @@ a:hover {
     -webkit-transition: color 0.2s;
 }
 
-footer {
-    height: 30px;
-    margin-top: 5rem;
-    margin-bottom: 8rem;
-}
-
-footer p {
-    font-size: 10px;
-    line-height: 12px;
-    text-align: center;
-}
-
 /* @media screen and (max-width: 800px) {
   header img {
     width: 25.7rem;
