+++ frontend/public/index.html
@@ -30,13 +30,6 @@
   <main id="root"></main>
 
   <div id="modal"></div>
-
-  <footer>
-    <p>
-      Copyright © 2022 Paolo Donadeo.<br />Rilasciato sotto licenza MIT, vedi
-      <a href="https://github.com/pdonadeo/e-uscito-joypad/blob/main/LICENSE">LICENSE.md</a>
-    </p>
-  </footer>
 </body>
 
 </html>
