+++ frontend/src/components/Footer.module.css
@@ -1,0 +1,19 @@
+footer {
+    height: 30px;
+    margin-top: 5rem;
+    margin-bottom: 8rem;
+    font-size: 10px;
+    line-height: 12px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+}
+
+footer a {
+    margin-left: .5em;
+}
+
+.bottomIcon {
+    height: 2em;
+}
